+++ pom.xml
@@ -289,6 +289,11 @@
 			<version>0.9.0</version>
 			<scope>test</scope>
 		</dependency>
+		<!-- WebJars Support -->
+		<dependency>
+			<groupId>org.webjars</groupId>
+			<artifactId>webjars-locator-core</artifactId>
+		</dependency>
 	</dependencies>
 
 	<build>
