+++ src/test/java/tqs/sparkflow/stationservice/service/OpenChargeMapServiceIT.java
@@ -20,19 +20,27 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 
 import tqs.sparkflow.stationservice.StationServiceApplication;
 import tqs.sparkflow.stationservice.TestcontainersConfiguration;
 import tqs.sparkflow.stationservice.config.TestConfig;
+import tqs.sparkflow.stationservice.config.OpenChargeMapTestConfig;
 import tqs.sparkflow.stationservice.repository.StationRepository;
 
 @SpringBootTest(
-    classes = {StationServiceApplication.class, TestConfig.class, TestcontainersConfiguration.class},
+    classes = {
+        StationServiceApplication.class,
+        TestConfig.class,
+        TestcontainersConfiguration.class,
+        OpenChargeMapTestConfig.class
+    },
     properties = {"spring.main.allow-bean-definition-overriding=true"})
 @ActiveProfiles("test")
 @ExtendWith(MockitoExtension.class)
+@TestPropertySource(properties = {"openchargemap.api.key=test-key"})
 class OpenChargeMapServiceIT {
 
   @Autowired
@@ -48,7 +56,6 @@ class OpenChargeMapServiceIT {
     stationRepository.deleteAll();
     openChargeMapService = new OpenChargeMapService(restTemplate, stationRepository, "test-key", "http://test-url");
   }
-
 
   @Test
   void whenPopulatingStationsWithInvalidCoordinates_thenThrowsException() {
@@ -92,4 +99,24 @@ class OpenChargeMapServiceIT {
         .hasMessage("Invalid Open Charge Map API key");
   }
 
+  @Test
+  void whenPopulatingStationsWithInvalidApiKey_thenThrowsException() {
+    // Given
+    double latitude = 38.7223;
+    double longitude = -9.1393;
+    int radius = 10;
+
+    // Mock RestTemplate to throw 401 Unauthorized
+    Mockito.lenient().when(restTemplate.exchange(
+            anyString(),
+            eq(HttpMethod.GET),
+            eq(null),
+            ArgumentMatchers.<ParameterizedTypeReference<List<Map<String, Object>>>>any()))
+        .thenThrow(new HttpClientErrorException(HttpStatus.UNAUTHORIZED));
+
+    // When/Then
+    assertThatThrownBy(() -> openChargeMapService.populateStations(latitude, longitude, radius))
+        .isInstanceOf(IllegalStateException.class)
+        .hasMessage("Access denied to Open Charge Map API");
+  }
 }
