+++ src/test/resources/application-test.properties
@@ -24,3 +24,7 @@ user.service.url=http://mock-user-service
 
 # Allow bean definition overriding
 spring.main.allow-bean-definition-overriding=true
+
+# Flyway configuration
+spring.flyway.enabled=false
+spring.flyway.baseline-on-migrate=false
